Adds contextual logging capability to cache wrapper (#22)
The memoing function that is returned by a `cacheWrapper` will now inspect the
arguments passed in by a caller. If any of the arguments appears to be a logger,
it will use that, otherwise it will fall back to the (global) logger passed at
construction time.

Why? This allows us to pass in contextual loggers, i.e., loggers that have some
context baked into them. In particular, we'll add a request id to the contextual
logger passed in with the calling context and be able to correlate cache lookups
with a per-service request context.